Syntax Reference

LanguageServer.LanguageServerInstance

LanguageServer.Range

LanguageServer.URI2

LanguageServer.get_line_offsets

LanguageServer.get_offset

LanguageServer.get_position_at

Main

LanguageServer.LanguageServerInstance — Type

LanguageServerInstance(pipe_in, pipe_out, env="", depot="", err_handler=nothing, symserver_store_path=nothing)

Construct an instance of the language server.

Once the instance is run, it will read JSON-RPC from pipe_out and write JSON-RPC from pipe_in according to the language server specification. For normal usage, the language server can be instantiated with LanguageServerInstance(stdin, stdout, false, "/path/to/environment").

Arguments

pipe_in::IO: Pipe to read JSON-RPC from.

pipe_out::IO: Pipe to write JSON-RPC to.

env::String: Path to the environment for which the language server is running. An empty string uses julia's default environment.

depot::String: Sets the JULIA_DEPOT_PATH where the language server looks for packages required in env.

err_handler::Union{Nothing,Function}: If not nothing, catch all errors and pass them to an error handler function with signature err_handler(err, bt). Mostly used for the VS Code crash reporting implementation.

symserver_store_path::Union{Nothing,String}: if nothing is passed, the symbol server cash is stored in a folder in the package. If an absolute path is passed, the symbol server will store the cache files in that path. The path must exist on disc before this is called.

LanguageServer.Range — Method

Range(Doc, rng)

Converts a byte offset range to a LSP Range.

LanguageServer.get_line_offsets — Function

get_line_offsets(doc::Document)

Updates the doc.lineoffsets field, an n length Array each entry of which gives the byte offset position of the start of each line. This always starts with 0 for the first line (even if empty).

LanguageServer.get_offset — Method

get_offset(doc, line, char)

Returns the byte offset position corresponding to a line/character position. This takes 0 based line/char inputs. Corresponding functions are available for Position and Range arguments, the latter returning a UnitRange{Int}.

LanguageServer.get_position_at — Method

get_position_at(doc, offset)

Returns the 0-based line and character position within a document of a given byte offset.

Base.run — Method

run(server::LanguageServerInstance)

Run the language server.

LanguageServer.URI2 — Type

A type that represents a URI. The main feature provided currently is that when the file:// scheme is used and the platform is Windows, equality and hashing is case-insensitive.

Requests

Protocol