Syntax Reference

LanguageServer.LanguageServerInstance

LanguageServer.Range

LanguageServer.URI2

LanguageServer.get_line_offsets

LanguageServer.get_offset

LanguageServer.get_position_at

Main

LanguageServer.LanguageServerInstance — Type

LanguageServerInstance(pipe_in, pipe_out, debug=false, env="", depot="")

Construct an instance of the language server.

Once the instance is run, it will read JSON-RPC from pipe_out and write JSON-RPC from pipe_in according to the language server specification. For normal usage, the language server can be instantiated with LanguageServerInstance(stdin, stdout, false, "/path/to/environment").

Arguments

pipe_in::IO: Pipe to read JSON-RPC from.

pipe_out::IO: Pipe to write JSON-RPC to.

debug::Bool: Whether to log debugging information with Base.CoreLogging.

env::String: Path to the environment for which the language server is running. An empty string uses julia's default environment.

depot::String: Sets the JULIA_DEPOT_PATH where the language server looks for packages required in env.

LanguageServer.Range — Method

Range(Doc, rng)

Converts a byte offset range to a LSP Range.

LanguageServer.get_line_offsets — Function

get_line_offsets(doc::Document)

Updates the doc.lineoffsets field, an n length Array each entry of which gives the byte offset position of the start of each line. This always starts with 0 for the first line (even if empty).

LanguageServer.get_offset — Method

get_offset(doc, line, char)

Returns the byte offset position corresponding to a line/character position. This takes 0 based line/char inputs. Corresponding functions are available for Position and Range arguments, the latter returning a UnitRange{Int}.

LanguageServer.get_position_at — Method

get_position_at(doc, offset)

Returns the 0-based line and character position within a document of a given byte offset.

Base.run — Method

run(server::LanguageServerInstance)

Run the language server.

LanguageServer.URI2 — Type

A type that represents a URI. The main feature provided currently is that when the file:// scheme is used and the platform is Windows, equality and hashing is case-insensitive.

Requests

Protocol