Syntax Reference

LanguageServer.LanguageServerInstance

LanguageServer.Range

LanguageServer.URI2

LanguageServer.get_line_offsets

LanguageServer.get_offset

LanguageServer.get_position_at

LanguageServer.is_completion_match

LanguageServer.is_diag_dependent_on_env

LanguageServer.runserver

Main

LanguageServer.Range — Method

Range(Doc, rng)

Converts a byte offset range to a LSP Range.

LanguageServer.get_line_offsets — Function

get_line_offsets(doc::Document)

Updates the doc.lineoffsets field, an n length Array each entry of which gives the byte offset position of the start of each line. This always starts with 0 for the first line (even if empty).

LanguageServer.get_offset — Method

get_offset(doc, line, char)

Returns the 0 based byte offset position corresponding to a line/character position. This takes 0 based line/char inputs. Corresponding functions are available for Position and Range arguments, the latter returning a UnitRange{Int}.

LanguageServer.get_position_at — Method

get_position_at(doc, offset)

Returns the 0-based line and character position within a document of a given byte offset.

LanguageServer.is_diag_dependent_on_env — Method

isdiagdependentonenv(diag::Diagnostic)::Bool

Is this diagnostic reliant on the current environment being accurately represented?

LanguageServer.LanguageServerInstance — Type

LanguageServerInstance(pipe_in, pipe_out, env="", depot="", err_handler=nothing, symserver_store_path=nothing)

Construct an instance of the language server.

Once the instance is run, it will read JSON-RPC from pipe_out and write JSON-RPC from pipe_in according to the language server specification. For normal usage, the language server can be instantiated with LanguageServerInstance(stdin, stdout, false, "/path/to/environment").

Arguments

pipe_in::IO: Pipe to read JSON-RPC from.

pipe_out::IO: Pipe to write JSON-RPC to.

env::String: Path to the environment for which the language server is running. An empty string uses julia's default environment.

depot::String: Sets the JULIA_DEPOT_PATH where the language server looks for packages required in env.

err_handler::Union{Nothing,Function}: If not nothing, catch all errors and pass them to an error handler function with signature err_handler(err, bt). Mostly used for the VS Code crash reporting implementation.

symserver_store_path::Union{Nothing,String}: if nothing is passed, the symbol server cash is stored in a folder in the package. If an absolute path is passed, the symbol server will store the cache files in that path. The path must exist on disc before this is called.

Base.run — Method

run(server::LanguageServerInstance)

Run the language server.

LanguageServer.runserver — Function

runserver(pipe_in=stdin, pipe_out=stdout[, env_path])

Run a LanguageServerInstance reading from pipe_in and writing to pipe_out.

The same options can be passed to runserver as to LanguageServerInstance. If env_path is not specified, attempt to pick an environment by considering in order of priority:

ARGS[1]: the first command-line argument passed to the invocation of julia.

The Julia project containing pwd().

The default Julia environment withing .julia/environments/v#.#.

Examples

The following invocation of Julia would set env_path to /home/example/repos/Example.jl:

julia --project=/path/to/LanguageServer.jl \
  -e "using LanguageServer, SymbolServer; runserver()" \
  /home/example/repos/Example.jl

Note

Due to a current bug, SymbolServer must be imported into Main.

If there was a Project.toml or JuliaProject.toml in /home/example/repos/Example.jl/, the following invocation would set env_path to /home/example/repos/Example.jl/; otherwise it would be set to .julia/environments/v#.# where v#.# is the major/minor version of Julia being invoked.

julia --project=/path/to/LanguageServer.jl \
  -e "using LanguageServer, SymbolServer; runserver()"

LanguageServer.URI2 — Type

A type that represents a URI. The main feature provided currently is that when the file:// scheme is used and the platform is Windows, equality and hashing is case-insensitive.

Requests

LanguageServer.is_completion_match — Function

is_completion_match(s::AbstractString, prefix::AbstractString, cutoff=3)

Returns true if s starts with prefix or has a sufficiently high fuzzy score.

Protocol